Moved status interfaces
