+++ client/src/interfaces/status/agreementStatus.ts
@@ -1,0 +1,4 @@
+export interface AgreementStatus {
+	name: string;
+	id: number;
+}
