+++ client/src/interfaces/status/reservationStatus.ts
@@ -1,0 +1,4 @@
+export interface ReservationStatus {
+	name: string;
+	id: number;
+}
