+++ client/src/interfaces/status/vehicleStatus.ts
@@ -1,0 +1,4 @@
+export interface VehicleStatus {
+	name: string;
+	id: number;
+}
