--- client/src/interfaces/statuses.ts
@@ -1,14 +1,0 @@
-export interface ReservationStatus {
-	name: string;
-	id: number;
-}
-
-export interface AgreementStatus {
-	name: string;
-	id: number;
-}
-
-export interface VehicleStatus {
-	name: string;
-	id: number;
-}
